requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
httpx
python-dotenv
llama-index
llama-index-embeddings-fastembed
llama-index-vector-stores-qdrant
llama-parse
groq
qdrant-client
Pillow
supabase
requests
numpy
pydantic
pyyaml
